Code cleanup

+++ src/Engine/Zoom.cpp
@@ -401,7 +401,7 @@ static int zoomSurface2X_XAxis_32bit(SDL_Surface *src, SDL_Surface *dst)
 					if (dsty + j >= dst->h) break;
 
 					*(pixelDst + (dst->pitch/sizeof(Uint32))*j) = dataDst;
-				} while(say[dsty + ++j] == 0); // fill in all relevant rows
+				} while (say[dsty + ++j] == 0); // fill in all relevant rows
 				
 				dataSrc >>= 16;
 				pixelDst++; // forward 4 bytes!
@@ -489,7 +489,7 @@ static int zoomSurface4X_XAxis_32bit(SDL_Surface *src, SDL_Surface *dst)
 					if (dsty + j >= dst->h) break;
 
 					*(pixelDst + (dst->pitch/sizeof(Uint32))*j) = dataDst;
-				} while(say[dsty + ++j] == 0); // fill in all relevant rows
+				} while (say[dsty + ++j] == 0); // fill in all relevant rows
 				
 				dataSrc >>= 8;
 				pixelDst++; // forward 4 bytes!
@@ -726,7 +726,7 @@ int Zoom::_zoomSurfaceY(SDL_Surface * src, SDL_Surface * dst, int flipx, int fli
 			// check the resolution to see which scale we need
 			for (size_t factor = 2; factor <= 5; factor++)
 			{
-				if (dst->w == src->w * factor && dst->h == src->h * factor)
+				if (dst->w == src->w * (int)factor && dst->h == src->h * (int)factor)
 				{
 					xbrz::scale(factor, (uint32_t*)src->pixels, (uint32_t*)dst->pixels, src->w, src->h);
 					return 0;
@@ -771,7 +771,7 @@ int Zoom::_zoomSurfaceY(SDL_Surface * src, SDL_Surface * dst, int flipx, int fli
 		// check the resolution to see which of scale2x, scale3x, etc. we need
 		for (size_t factor = 2; factor <= 4; factor++)
 		{
-			if (dst->w == src->w * factor && dst->h == src->h * factor && !scale_precondition(factor, src->format->BytesPerPixel, src->w, src->h))
+			if (dst->w == src->w * (int)factor && dst->h == src->h * (int)factor && !scale_precondition(factor, src->format->BytesPerPixel, src->w, src->h))
 			{
 				scale(factor, dst->pixels, dst->pitch, src->pixels, src->pitch, src->format->BytesPerPixel, src->w, src->h);
 				return 0;
